add "bonus" can transceiver ;)
this is so we can use the round 1 protos for another project...

also: replace srv05 with an alternate, also jlc preferred, but cheaper
part.  Update power budgets.

We've done some initial work with the pcb, but we're running into weird
kicad wx bugs, so we're going to save it all here, and commit, and
perhaps continue with 7.99.....